requirements-dev.txt
black
flake8
ipdb
ipython
isort
pytest >=5.2
pytest-flask
pytest-cov